core: Move buffer_pool.rs to storage module
