+++ core/storage/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod btree;
+pub(crate) mod buffer_pool;
 pub(crate) mod pager;
 pub(crate) mod sqlite3_ondisk;
 pub(crate) mod wal;
