+++ core/storage/pager.rs
@@ -1,4 +1,4 @@
-use crate::buffer_pool::BufferPool;
+use crate::storage::buffer_pool::BufferPool;
 use crate::storage::sqlite3_ondisk::{self, DatabaseHeader, PageContent};
 use crate::storage::wal::Wal;
 use crate::{Buffer, PageSource, Result};
